+++ headers/frame.hh
@@ -6,4 +6,4 @@ void run(App* app);
 extern PlayerControls player;
 extern f64 incCounter;
 extern f64 fov;
-extern f32 x, y, z;
+extern f64 x, y, z;
